Initial commit
A bit f*cked right now, but I am in the process of fixing it.  I am removing Threading for the time being, to enable me in debugging this crap.   But once I have it thread-safe, I plan on re-enabling threading.  I hope that I am getting threading right, and JCIP and I are going to be constant companions for a while.